requirements.txt
Flask>=1.1.2
requests>=2.25.1
python-dotenv>=0.19.2
pytest>=6.2.5
